EHN: Heuristically infer midpoint provenance when missing
This addresses issue #19.

Some fiducial lists, when attached to a model (mesh), do not have any of
the JSON data MRML node attributes that are conventionally used by Q3DC
and related modules. This can happen when the fiducial list is loaded as
a fcsv instead of in a MRML file. (Can a fiducial list ever be in a MRML
file? Does this preserve the JSON data?) The fcsv data *does* contain
the field called "sel", which is short for "selected". Midpoints created
by Q3DC are initially deselected, as a hack to cause their color to
indicate their role. Thus the fcsv indirectly indicates that certain
points were constructed as midpoints.

The recover_midpoint_provenance method uses this information to guess
which points were used to construct which midpoints. It is assumed that
any selected point is not a midpoint. For each deselected point, x, a
search is performed for a pair of selected points whose midpoint is near
x. For each selected point, y, the point z on the opposite side of x
from y is calculated, and a kd-tree is consulted to determine if there
is a selected point near z. If so, y and z are accepted as the "parent"
points that were used to generate x.

Notice that it is assumed that no midpoint was constructed from any
other midpoint.

The kd-tree implementation from scipy.spatial is used, so this module
is installed with slicer.util.pip_install if needed.

If the "On Surface" check box is checked (the one next to the "Connected
Landmarks" selection box), then points are projected onto the attached
model (mesh). Using the inferred midpoint provenances, the midpoints are
now moved to the new midpoint of their inferred parents instead of being
independently projected onto the model.

numpy is now imported as np after common convention.