BUG: Ensure all fiducial points are in graph before toposort
issue #19

A topological sort is used to order the fiducial points before updating
their positions when changing the model associated with a fiducial list.
All the fiducial points must be represented in the directed graph so
that they are all in the result of the topological sort. This includes
fiducial points that are not a midpoint and are not a parent of any
midpoint. The easiest way to accomplish this is to add each point to the
graph unconditionally, and don't rely on inferring the node list from
the edge list.