EHN: Select non-midpoints that start selected during loading
issue #19

When a fiducial list is loaded as an fcsv file, a fiducial x may be
non-selected in the fscv. Q3DC's loading system will attempt to find a
pair of other fiducials whose midpoint is x. If it fails, it will begin
treating x as a regular non-midpoint fiducial. This change indicates
when this happens to the user by changing the selection status of x to
selected, or non-midpoint.