EHN: Harmonize UI text and CSV headers
The headers of CSV files written by Q3DC refer to landmark points by
names similar to what is displayed in the Q3DC UI. This commit brings
these into better agreement. In addition, the headers of linePoint and
angle CSV files is updated to reflect the actual content of those CSVs.