ENH: Simplify line angle computation
Use numpy to compute the signed angle between projected lines. This results in less repeated code, and the resulting computation looks more like the relevant linear algebra.

The plane to be projected is also parameterized in `axis`, so the same calculation function may be used for pitch, roll, and yaw. `axis` specifies the index to be removed from the vectors.